+++ src/constants.ts
@@ -1,9 +1,5 @@
 export const EXTENSION_NAME = "wipstream";
 
-export const LEGACY_REPOSITORY_CONFIG_VERSION = "1";
-export const REPOSITORY_CONFIG_VERSION = "2";
-
 export const CONFIG_KEYS = {
-  version: "wipstream.version",
   remote: "wipstream.remote",
 } as const;